requirements.txt
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
dataclasses==0.6
dj-database-url==2.3.0
Django==4.2.11
django-cors-headers==4.6.0
django-mongoengine==0.5.6
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djongo==1.2.31
dnspython==2.7.0
dotenv==0.9.9
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httptools==0.6.4
Hypercorn==0.17.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
lxml==6.0.0
mongoengine==0.29.1
msgpack==1.1.0
packaging==25.0
pillow==11.0.0
priority==2.0.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pymongo==3.11.4
pyOpenSSL==25.0.0
PyPDF2==3.0.1
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.4
rsa==4.9.1
service-identity==24.2.0
setuptools==75.8.0
sniffio==1.3.1
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.5.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.2
whitenoise==6.8.2
wsproto==1.2.0
zope.interface==7.2
google-auth==2.40.3
requests==2.32.4
